requirements.txt
# MediaCrawler 项目依赖包
# 基础框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# 浏览器自动化
playwright==1.40.0

# 数据库
aiomysql==0.2.0
pymysql==1.1.0

# 文件处理
aiofiles==23.2.1
pyyaml==6.0.1

# 网络请求
requests==2.31.0
aiohttp==3.9.1

# 数据解析
beautifulsoup4==4.12.2
lxml==4.9.3

# 异步支持
asyncio-throttle==1.0.2

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# 日志
loguru==0.7.2

# 时间处理
python-dateutil==2.8.2

# 正则表达式增强
regex==2023.10.3

# 进度条
tqdm==4.66.1

# 其他依赖
httpx==0.24.0
Pillow==9.5.0
tenacity==8.2.2
redis~=4.6.0
python-dotenv==1.0.1
jieba==0.42.1
wordcloud==1.9.3
matplotlib==3.9.0
parsel==1.9.1
pyexecjs==1.5.1
pandas==2.2.3

# 存储系统相关依赖
minio==7.2.0
sqlalchemy==2.0.25
python-multipart==0.0.6

# 图像处理
opencv-python-headless==4.8.1.78

# 调度任务
apscheduler==3.10.4